requirements.txt
psutil
python-dotenv
PyAutoGUI
ifcfg